requirements.txt
aiofiles==24.1.0
aiogram==3.20.0.post0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.2
django-cleanup==9.0.0
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-nested-routers==0.94.1
drf-yasg==1.21.10
eval_type_backport==0.2.2
frozenlist==1.6.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
inflection==0.5.1
magic-filter==1.0.12
mistralai==1.6.0
multidict==6.4.3
packaging==24.2
pillow==11.1.0
propcache==0.3.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
pydantic==2.11.3
pydantic_core==2.33.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
text-unidecode==1.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
yarl==1.20.0

celery 
redis
